+++ src/components/Filter/components/MultiFilter/MultiFilter.style.ts
@@ -26,23 +26,4 @@ export const optionsWrapper = () => (theme: Theme) => {
   `;
 };
 
-export const menuStyle = () => (theme: Theme) => {
-  const borderConfig = textInputConfig.types[theme.colorScheme].outlined.border;
 
-  return css`
-    position: absolute;
-    top: ${rem(48)};
-    min-width: ${rem(280)};
-    left: 0;
-    height: auto;
-    border: ${rem(borderConfig.width)} solid
-      ${theme.utils.getColor(borderConfig.color.default.name, borderConfig.color.default.shade)};
-    border-radius: ${theme.spacing.xsm};
-    background-color: ${theme.palette.white};
-    box-shadow: ${theme.elevation['02']};
-    z-index: 500;
-    overflow: hidden;
-    min-width: 100%;
-    max-width: ${rem(440)};
-  `;
-};
